pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "histoRAG"
version = "0.0.1"
description = "Retrieval-Augmented Histopathology Atlas — FAU BIMAP SS2026"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "open_clip_torch>=2.24",
    "timm>=0.9.8",
    "openslide-bin>=4.0",
    "openslide-python>=1.3",
    "faiss-cpu>=1.7",
    "numpy",
    "pandas",
    "pyarrow",
    "Pillow",
    "streamlit>=1.30",
    "pyyaml",
    "tqdm",
    "scikit-learn",
    "huggingface_hub",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-cov",
    "ruff",
    "black",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
